Assorted memory leak and error throwing fixes
1. Adds an isTempId method for easy testing to see if we're dealing with a local/unsynced resource. Sadly, using the syncState was too cumbersome for this.
2. Side-effects in resetting all queries should prevent errors
3. Conversation._createResult and _updateUnreadCountEvent firing after destroying the resource will now be ignored
4. Conversation.delete will ignore if it has a temporary ID
5. Messages and Conversations that fail to load will now be fully destroyed not just removed but left around.
6. System bus unsubscribe is now optimized
7. sync-event firing property should never get stuck
8. sync-manager not supports the not-found error
9. Switches from dates to times in sync-manager.

+++ src/conversation.js
@@ -284,6 +284,7 @@ class Conversation extends Syncable {
    * @param  {Object} result
    */
   _createResult({ success, data }) {
+    if (this.isDestroyed) return;
     if (success) {
       this._createSuccess(data);
     } else if (data.id === 'conflict') {
@@ -532,13 +533,15 @@ class Conversation extends Syncable {
    */
   delete(destroy) {
     const id = this.id;
-    const client = this.getClient();
-    this._xhr({
-      method: 'DELETE',
-      url: '?destroy=' + Boolean(destroy),
-    }, result => {
-      if (!result.success) Conversation.load(id, client);
-    });
+    if (!this.isTempId()) {
+      const client = this.getClient();
+      this._xhr({
+        method: 'DELETE',
+        url: '?destroy=' + Boolean(destroy),
+      }, result => {
+        if (!result.success) Conversation.load(id, client);
+      });
+    }
 
     this._deleted();
     this.destroy();
@@ -870,7 +873,7 @@ class Conversation extends Syncable {
     if (!result.success) {
       this.syncState = Constants.SYNC_STATE.NEW;
       this.trigger('conversations:loaded-error', { error: result.data });
-      this.getClient()._removeConversation(this);
+      this.destroy();
     } else {
       // If successful, copy the properties into this object
       this._populateFromServer(result.data);
@@ -927,6 +930,9 @@ class Conversation extends Syncable {
    * Any triggering of this from a websocket patch unread_message_count should wait a second before firing any events
    * so that if there are a series of these updates, we don't see a lot of jitter.
    *
+   * NOTE: _oldUnreadCount is used to pass data to _updateUnreadCountEvent because this method can be called many times
+   * a second, and we only want to trigger this with a summary of changes rather than each individual change.
+   *
    * @method __updateUnreadCount
    * @private
    * @param  {number} newValue
@@ -949,6 +955,7 @@ class Conversation extends Syncable {
    * @private
    */
   _updateUnreadCountEvent() {
+    if (this.isDestroyed) return;
     const oldValue = this._oldUnreadCount;
     const newValue = this.__unreadCount;
     this._oldUnreadCount = undefined;
